requirements.txt
urwid
pyoath-toolkit
python-gnupg
